Fix: Conduction/Valence band level, and nie calculation under Bandgap narrowing
There is an error in the Medici manual regarding the Nie calculation in the presence of bandgap narrowing.
I fixed it for the case of Silicon.

Also, the Ec/Ev calculation in semiequ1e.cc seems not so right.
It is obvious that Ec-Ev =\= Eg.
In fact, the Eg value stored in aux[] has already taken BGN into account.

The formula stated in your user manual are correct, and agree with ISE manual. I coded accordingly.
I didn't really check for the case with Fermi-dirac statistics, though.

Signed-off-by: Shen Chen <shenchen@nus.edu.sg>